start with auth.js

+++ Auth/Practice/app/api/todos/[id]/route.js
@@ -4,6 +4,9 @@ import { connectDB } from "@/lib/connectDB";
 export async function GET(_, { params }) {
   await connectDB();
 
+  const myCookie = await cookies();
+  const userId = myCookie.get("userId")?.value;
+
   const { id } = await params;
   const todo = await Todo.findById(id);
   if (!todo) {
@@ -19,6 +22,19 @@ export async function GET(_, { params }) {
 
 export async function PUT(request, { params }) {
   await connectDB();
+
+  const myCookie = await cookies();
+  const userId = myCookie.get("userId")?.value;
+
+  if (!userId) {
+    return Response.json(
+      { error: "You are not authorized to edit this todo" },
+      {
+        status: 401,
+      }
+    );
+  }
+
   const editTodoData = await request.json();
   const { id } = await params;
   const editedTodo = await Todo.findByIdAndUpdate(id, editTodoData, {
@@ -30,6 +46,16 @@ export async function PUT(request, { params }) {
 
 export async function DELETE(_, { params }) {
   await connectDB();
+  const myCookie = await cookies();
+  const userId = myCookie.get("userId")?.value;
+  if (!userId) {
+    return Response.json(
+      { error: "You are not authorized to delete this todo" },
+      {
+        status: 401,
+      }
+    );
+  }
   const { id } = await params;
   await Todo.findByIdAndDelete(id);
   return new Response(null, {
